refactor(tests): split popup.test.ts into focused test modules
Replace the single 675-line popup.test.ts with three focused files:
popup.display (getScoreColor, displayCategoryContent, displayResults),
popup.ui (switchTab, updatePageInfo, showLoadingState, showError), and
popup.actions (exportResults, runAnalysis). Shared DOM helpers
(setupPopupDOM, buildCategory, buildResult) live in tests/helpers/popup.ts.
